vton_project/requirements.txt
torch
torchvision
opencv-python
numpy
pillow
tqdm
lpips
matplotlib
